requirements.txt
langchain==0.1.5
langchainhub==0.1.14
langchain_openai==0.0.5
langchain-core==0.1.18
argparse==1.4.0
python-dotenv==1.0.0
chromadb==0.4.22
langchain-community==0.0.17
